pyproject.toml
[tool.poetry]
name = "dataiku-mlops"
version = "0.1.0"
description = "MLOps API Wrapper for Dataiku Data Science Studio"
license = "MIT"
authors = ["Aniket Mishrikotkar <aniketmishrikotkar@gmail.com>"]
maintainers = ["Aniket Mishrikotkar <aniketmishrikotkar@gmail.com>"]
readme = "README.md"
homepage = ""
repository = ""

[tool.poetry.dependencies]
python = "^3.10"
dataiku-api-client = "^12.2.3"
pendulum = "^2.1.2"
loguru = "^0.7.2"
pre-commit = "^3.5.0"
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
